backend/requirements.txt
cffi==1.2.1
cryptography==1.0.2
Django==1.8.5
djangorestframework==3.2.4
djangorestframework-jwt==1.7.2
idna==2.0
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
pyOpenSSL==0.15.1
six==1.10.0
